requirements.txt
ray
PyYAML==5.3.1
python-dotenv
langchain
llama-index
pypdf
faiss-cpu
tiktoken